Chore: coupon-preview-item 이름 변경 및 type 파일 분리
