+++ src/components/coupon-registration/coupon-preview/coupoin-preview-item/type.ts
@@ -1,0 +1,8 @@
+export type CouponPreviewItemProps = {
+  data: {
+    roomName: string;
+    couponName: string;
+    couponPrice: number;
+    couponAmount: number;
+  };
+};
